unit tests for Controller, added some cases and fixed checkstyle mistakes
