+++ src/main/java/com/example/usermanagementapi/config/MyMapperConfig.java
@@ -1,11 +1,12 @@
 package com.example.usermanagementapi.config;
 
+import org.mapstruct.MapperConfig;
 import org.mapstruct.NullValueCheckStrategy;
 
-@org.mapstruct.MapperConfig(
+@MapperConfig(
         componentModel = "spring",
         nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
         implementationPackage = "<PACKAGE_NAME>.impl"
 )
-public interface MapperConfig {
+public interface MyMapperConfig {
 }
