+++ src/main/java/com/example/usermanagementapi/mapper/UserRequestDtoMapper.java
@@ -1,11 +1,11 @@
 package com.example.usermanagementapi.mapper;
 
-import com.example.usermanagementapi.config.MapperConfig;
+import com.example.usermanagementapi.config.MyMapperConfig;
 import com.example.usermanagementapi.dto.UserRequestDto;
 import com.example.usermanagementapi.model.User;
 import org.mapstruct.Mapper;
 
-@Mapper(config = MapperConfig.class)
+@Mapper(config = MyMapperConfig.class)
 public interface UserRequestDtoMapper {
     User mapToModel(UserRequestDto userRequestDto);
 }
