+++ src/main/java/com/example/usermanagementapi/repository/UserRepository.java
@@ -3,7 +3,6 @@ package com.example.usermanagementapi.repository;
 import com.example.usermanagementapi.model.User;
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
